--- a/include/rocksdb/iterator.h
+++ b/include/rocksdb/iterator.h
@@ -37,6 +37,8 @@ class Cleanable {
   typedef void (*CleanupFunction)(void* arg1, void* arg2);
   void RegisterCleanup(CleanupFunction function, void* arg1, void* arg2);
   void DelegateCleanupsTo(Cleanable* other);
+  // DoCleanup and also resets the pointers for reuse
+  inline void Reset();
 
  protected:
   struct Cleanup {
@@ -48,6 +50,11 @@ class Cleanable {
   Cleanup cleanup_;
   // It also becomes the owner of c
   void RegisterCleanup(Cleanup* c);
+
+ private:
+  // Performs all the cleanups. It does not reset the pointers. Making it
+  // private to prevent misuse
+  inline void DoCleanup();
 };
 
 class Iterator : public Cleanable {