--- a/include/rocksdb/iterator.h
+++ b/include/rocksdb/iterator.h
@@ -38,7 +38,7 @@ class Cleanable {
   void RegisterCleanup(CleanupFunction function, void* arg1, void* arg2);
   void DelegateCleanupsTo(Cleanable* other);
   // DoCleanup and also resets the pointers for reuse
-  inline void Reset();
+  void Reset();
 
  protected:
   struct Cleanup {
@@ -54,7 +54,17 @@ class Cleanable {
  private:
   // Performs all the cleanups. It does not reset the pointers. Making it
   // private to prevent misuse
-  inline void DoCleanup();
+  inline void DoCleanup() {
+    if (cleanup_.function != nullptr) {
+      (*cleanup_.function)(cleanup_.arg1, cleanup_.arg2);
+      for (Cleanup* c = cleanup_.next; c != nullptr;) {
+        (*c->function)(c->arg1, c->arg2);
+        Cleanup* next = c->next;
+        delete c;
+        c = next;
+      }
+    }
+  }
 };
 
 class Iterator : public Cleanable {