@maysamyabandeh 2016-12-22
The sandcastle job fails with this error:
```
make: *** No rule to make target `table/cleanable_test.o', needed by `cleanable_test'.  Stop.
```
Not sure what is the problem since it complies fine in my dev machine and finds the implicit rule as for other objects in the Makefile. Here is the change to the Makefile:
```
diff --git a/Makefile b/Makefile
index d768647..11220ff 100644
--- a/Makefile
+++ b/Makefile
@@ -324,6 +324,7 @@ TESTS = \
        db_properties_test \
        db_table_properties_test \
        autovector_test \
+       cleanable_test \
        column_family_test \
        table_properties_collector_test \
        arena_test \
@@ -1147,6 +1148,9 @@ full_filter_block_test: table/full_filter_block_test.o $(LIBOBJECTS) $(TESTHARNE
 log_test: db/log_test.o $(LIBOBJECTS) $(TESTHARNESS)
        $(AM_LINK)
 
+cleanable_test: table/cleanable_test.o $(LIBOBJECTS) $(TESTHARNESS)
+       $(AM_LINK)
+
 table_test: table/table_test.o $(LIBOBJECTS) $(TESTHARNESS)
        $(AM_LINK)
 ```

@IslamAbdelRahman any idea?