@maysamyabandeh 2016-12-23
The sandcastle tests failures seems false positives. The core dump shows the following:
```
Core was generated by `./external_sst_file_test --gtest_filter=ExternalSSTFileTest.PickedLevelDynamic'.

(gdb) bt
#0  0x00002ab31ceb7cfa in  ()
#1  0x00000000005283a5 in __gnu_cxx::__exchange_and_add_single (__val=-1, __mem=0x7ffc1ce68714)
    at /mnt/gvfs/third-party2/gcc/cf7d14c625ce30bae1a4661c2319c5a283e4dd22/4.9.x/centos6-native/108cf83/include/c++/4.9.x-google/ext/atomicity.h:68
#2  0x00000000005283a5 in __gnu_cxx::__exchange_and_add_dispatch (__val=-1, __mem=0x7ffc1ce68714)
    at /mnt/gvfs/third-party2/gcc/cf7d14c625ce30bae1a4661c2319c5a283e4dd22/4.9.x/centos6-native/108cf83/include/c++/4.9.x-google/ext/atomicity.h:84
#3  0x00000000005283a5 in std::_Sp_counted_base<(__gnu_cxx::_Lock_policy)2>::_M_release() (this=0x7ffc1ce68708)
    at /mnt/gvfs/third-party2/gcc/cf7d14c625ce30bae1a4661c2319c5a283e4dd22/4.9.x/centos6-native/108cf83/include/c++/4.9.x-google/bits/shared_ptr_base.h:163
#4  0x00000000005283a5 in std::__shared_count<(__gnu_cxx::_Lock_policy)2>::~__shared_count() (this=0x7ffc1ce68288, __in_chrg=<optimized out>)
    at /mnt/gvfs/third-party2/gcc/cf7d14c625ce30bae1a4661c2319c5a283e4dd22/4.9.x/centos6-native/108cf83/include/c++/4.9.x-google/bits/shared_ptr_base.h:666
#5  0x00000000005283a5 in std::__shared_ptr<rocksdb::FilterPolicy const, (__gnu_cxx::_Lock_policy)2>::~__shared_ptr() (this=0x7ffc1ce68280, __in_chrg=<optimized out>)
    at /mnt/gvfs/third-party2/gcc/cf7d14c625ce30bae1a4661c2319c5a283e4dd22/4.9.x/centos6-native/108cf83/include/c++/4.9.x-google/bits/shared_ptr_base.h:914
#6  0x00000000005283a5 in std::shared_ptr<rocksdb::FilterPolicy const>::~shared_ptr() (this=0x7ffc1ce68280, __in_chrg=<optimized out>)
    at /mnt/gvfs/third-party2/gcc/cf7d14c625ce30bae1a4661c2319c5a283e4dd22/4.9.x/centos6-native/108cf83/include/c++/4.9.x-google/bits/shared_ptr.h:93
#7  0x00000000005283a5 in rocksdb::anon::OptionsOverride::~OptionsOverride() (this=0x7ffc1ce68280, __in_chrg=<optimized out>) at ./db/db_test_util.h:114
#8  0x00000000005283a5 in rocksdb::DBTestBase::OptionsForLogIterTest() (this=<optimized out>) at db/db_test_util.cc:951
#9  0x00000000005285ac in rocksdb::DBTestBase::ChangeCompactOptions() (this=0x7ffc1ce68700) at db/db_test_util.cc:166
#10 0x0000000000000000 in  ()
```
which starts at this line
```
166         auto options = CurrentOptions();
```
Which seems irrelevant to the changes made by this patch. Moreover it passed when run locally:
```
[?0] myabandeh@devvm17154:~/rocksdb[cleanabledelagate)]$ ./external_sst_file_test --gtest_filter=ExternalSSTFileTest.PickedLevelDynamic
Note: Google Test filter = ExternalSSTFileTest.PickedLevelDynamic
[==========] Running 1 test from 1 test case.
[----------] Global test environment set-up.
[----------] 1 test from ExternalSSTFileTest
[ RUN      ] ExternalSSTFileTest.PickedLevelDynamic
[       OK ] ExternalSSTFileTest.PickedLevelDynamic (10 ms)
[----------] 1 test from ExternalSSTFileTest (10 ms total)

[----------] Global test environment tear-down
[==========] 1 test from 1 test case ran. (10 ms total)
[  PASSED  ] 1 test.
```